Split /traces routes in different files.
