+++ app/index.js
@@ -33,7 +33,7 @@ export default async function () {
   await server.register({
     plugin: hapiRouter,
     options: {
-      routes: 'app/routes/*.js'
+      routes: ['app/routes/*.js', 'app/routes/traces/*.js']
     }
   });
