+++ app/routes/traces/post.js
@@ -1,11 +1,8 @@
-import Joi from '@hapi/joi';
-import traces from '../models/traces';
-import logger from '../services/logger';
-import db from '../services/db';
 import Boom from '@hapi/boom';
-import config from 'config';
-
-const idLength = config.get('idLength');
+import Joi from '@hapi/joi';
+import db from '../../services/db';
+import logger from '../../services/logger';
+import traces from '../../models/traces';
 
 /**
  * @api {post} /traces POST
@@ -27,7 +24,7 @@ const idLength = config.get('idLength');
  *      "message": "Oops!"
  *     }
  */
-module.exports = [
+export default [
   {
     path: '/traces',
     method: ['POST'],
@@ -104,89 +101,6 @@ module.exports = [
 
           // Return as TraceJSON
           return traces.asTraceJson(trace);
-        } catch (error) {
-          logger.error(error);
-          return Boom.badImplementation('Unexpected error.');
-        }
-      }
-    }
-  },
-  {
-    path: '/traces/{id}',
-    method: ['PATCH'],
-    options: {
-      auth: 'jwt',
-      validate: {
-        params: Joi.object({
-          id: Joi.string().length(idLength)
-        }),
-        payload: Joi.object({
-          description: Joi.string()
-        })
-      },
-      handler: async function (request) {
-        try {
-          // Get trace
-          const { id } = request.params;
-          const [trace] = await traces.get(id);
-
-          // Verify ownership
-          const { osmId, isAdmin } = request.auth.credentials;
-          if (trace.ownerId !== osmId && !isAdmin) {
-            return Boom.forbidden('Must be owner or admin to edit a trace.');
-          }
-
-          // Patch
-          const { description } = request.payload;
-          const [patchedTrace] = await traces
-            .update(id, { description })
-            .returning([
-              'id',
-              'ownerId',
-              'description',
-              'length',
-              'recordedAt',
-              'uploadedAt',
-              'updatedAt',
-              'timestamps',
-              db.raw('ST_AsGeoJSON(geometry) as geometry')
-            ]);
-
-          // Return as TraceJSON
-          return traces.asTraceJson(patchedTrace);
-        } catch (error) {
-          logger.error(error);
-          return Boom.badImplementation('Unexpected error.');
-        }
-      }
-    }
-  },
-  {
-    path: '/traces/{id}',
-    method: ['DELETE'],
-    options: {
-      auth: 'jwt',
-      validate: {
-        params: Joi.object({
-          id: Joi.string().length(idLength)
-        })
-      },
-      handler: async function (request) {
-        try {
-          // Get trace
-          const { id } = request.params;
-          const [trace] = await traces.get(id);
-
-          // Verify ownership
-          const { osmId, isAdmin } = request.auth.credentials;
-          if (trace.ownerId !== osmId && !isAdmin) {
-            return Boom.forbidden('Must be owner or admin to edit a trace.');
-          }
-
-          // Perform delete
-          await traces.del(id);
-
-          return 'ok';
         } catch (error) {
           logger.error(error);
           return Boom.badImplementation('Unexpected error.');
